net-base: Prepare for base::span API change
The parameter types of some functions (e.g. subspan) are becoming
stricter, see https://crrev.com/c/6044946. Make the necessary
changes and also make more use of first() and compile-time sizes.

BUG=b:381490721
TEST=CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6057329
Auto-Submit: Georg Neis <neis@chromium.org>
Commit-Queue: Chuwei He <chuweih@google.com>
Tested-by: Georg Neis <neis@chromium.org>

+++ net-base/attribute_list.cc
@@ -93,14 +93,17 @@ bool AttributeList::IterateAttributes(
 
     base::span<const uint8_t> value;
     if (attribute->nla_len > NLA_HDRLEN) {
-      value = remaining.subspan(NLA_HDRLEN, attribute->nla_len - NLA_HDRLEN);
+      value = remaining.subspan(
+          static_cast<size_t>(NLA_HDRLEN),
+          static_cast<size_t>(attribute->nla_len - NLA_HDRLEN));
     }
     if (!method.Run(attribute->nla_type, value)) {
       return false;
     }
 
     if (remaining.size() >= NLA_ALIGN(attribute->nla_len)) {
-      remaining = remaining.subspan(NLA_ALIGN(attribute->nla_len));
+      remaining =
+          remaining.subspan(static_cast<size_t>(NLA_ALIGN(attribute->nla_len)));
     } else {
       remaining = {};
     }
